requirements.txt
# === Core Data & Math ===
pandas>=2.2.2
numpy>=1.26.4
pandas-ta>=0.4.67b0
yfinance>=0.2.50

# === Database ===
pymongo>=4.8.0
python-dotenv>=1.0.1

# === Optional for News & Fundamentals (future) ===
textblob>=0.18.0
nltk>=3.9.1

# === API / Dashboard (future expansion) ===
fastapi>=0.115.0
uvicorn>=0.30.6
pydantic>=2.9.2

# === Utilities ===
requests>=2.32.3
tqdm>=4.66.5
